requirements.txt
pyYAML==3.12
nose==1.3.7